Restructure PLC tests to mirror camera test organization
- Split monolithic test_plcs.py into focused test modules
- Move Allen Bradley PLC tests to backends/allen_bradley/test_mock_allen_bradley.py
- Move PLC manager tests to core/test_plc_manager.py
- Add shared async fixtures and utilities to conftest.py
- Update documentation to reflect new test structure
